Refactoring models and adding hash generation methods to user model
